cleanup static reference
remove static shared things in ViewController
define new singleton class NumswPlayground
change book initialize code
`_setupRender` to `NumswPlayground.initialize()`
