--- playbook/numsw-ipad.playgroundbook/Contents/Sources/playgrounds/PlaygroundSetup.swift
@@ -1,16 +1,0 @@
-//
-//  PlaygroundSetup.swift
-//  Charts
-//
-
-import UIKit
-import PlaygroundSupport
-
-
-public func _setupRender(){
-
-    let viewController:RenderViewController = RenderViewController()
-
-    PlaygroundPage.current.liveView = viewController as! UIViewController
-
-}
